// src/app/api/generated/models/orders-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface OrdersDto {
  customerID?: (string | null);
  employeeID?: (number | string | null);
  freight?: (number | string | null);
  orderDate?: (string | null);
  orderID: (number | string);
  requiredDate?: (string | null);
  shipAddress?: (string | null);
  shipCity?: (string | null);
  shipCountry?: (string | null);
  shipName?: (string | null);
  shipPostalCode?: (string | null);
  shipRegion?: (string | null);
  shipVia?: (number | string | null);
  shippedDate?: (string | null);
}